+++ config.js
@@ -1,7 +1,7 @@
 System.config({
-  baseURL: "",
-  defaultJSExtensions: true,
-  transpiler: "typescript",
+    baseURL: "",
+    defaultJSExtensions: true,
+    transpiler: "typescript",
   paths: {
     "npm:*": "jspm_packages/npm/*",
     "github:*": "jspm_packages/github/*"
